% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bc_log_ratio}
\alias{bc_log_ratio}
\title{Bias corrected log ratio and variance}
\usage{
bc_log_ratio(observations, phase, first_level_base = TRUE)
}
\arguments{
\item{observations}{vector of observations}

\item{phase}{factor or vector indicating levels of the PIR measurements.}

\item{first_level_base}{Logical value indicating if the first level of \code{phase} is the baseline level. Default is \code{TRUE}.}
}
\value{
A list containing two elements. The first element, \code{bias_corrected_R}, is the bias-corrected log ratio. The second element, \code{variance_R}, is the variance of the bias-corrected log ratio.
}
\description{
Estimates the bias corrected log ratio and the variance of the log ratio
}
\details{
The \code{observations} vector can be in any order corresponding to the factor or vector \code{phase}.
The levels of \code{phase} can be any two levels, such as "A" and "B", "base" and "treat", or "0" and "1".
If there are more than two levels in \code{phase} this function will not work.
By default it is assumed that the first level in \code{phase} is the baseline level,
but this can be changed by setting \code{first_level_base} to \code{FALSE}.
}
\examples{
#Get a log ratio and variance estimate for Carl from Moes dataset
data(Moes)
with(subset(Moes, Case == "Carl"),
 bc_log_ratio(observations = outcome, phase = Phase,
 first_level_base = FALSE))
}

