% Generated by roxygen2 (4.0.1): do not edit by hand
\name{continuous_duration_recording}
\alias{continuous_duration_recording}
\title{Applies continuous duration recording to a behavior stream}
\usage{
continuous_duration_recording(BS)
}
\arguments{
\item{BS}{object of class \code{behavior_stream}}
}
\value{
Vector of proportions.
}
\description{
Calculates the proportion of session time during which behavior occurs.
}
\examples{
BS <- r_behavior_stream(n = 5, mu = 3, lambda = 10,
                       F_event = F_exp(), F_interim = F_exp(), stream_length = 100)
continuous_duration_recording(BS)
}

