% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{Dunlap}
\alias{Dunlap}
\title{Dunlap et al.(1994) data}
\format{A data frame with 58 observations on 7 variables

\itemize{
\item [,1] \code{Case} The participant for whom the observation took place
\item [,2] \code{Phase} The level of the observation ("Choice" vs. "No Choice")
\item [,3] \code{Session} The observation session # for each participant
\item [,4] \code{outcome} The summary PIR measurement for the observation session
\item [,5] \code{active_length} The length of the active observation interval, in seconds
\item [,6] \code{rest_length} The length of the recording interval, in seconds
\item [,7] \code{intervals} The total number of intervals in the observation session
}}
\usage{
Dunlap
}
\description{
Single case design data measured with partial interval recording from a study of the effect of providing Choice between academic activities on the disruptive behavior of three elementary school students with emotional and behavioral disorders. For this data "No Choice" is the baseline phase. Data were extracted from the figures in the publicaton.
}
\references{
Dunlap, G., DePerczel, M., Clarke, S., Wilson, D., Wright, S., White, R., & Gomez, A. (1994). Choice making to promote adaptive behavior for students with emotional and behavioral challenges. Journal of Applied Behavior Analysis, 27 (3), 505-518.
}
\keyword{datasets}

