% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/observation_method_filters.R
\name{event_counting}
\alias{event_counting}
\title{Applies event counting to a behavior stream}
\usage{
event_counting(BS)
}
\arguments{
\item{BS}{object of class \code{behavior_stream}}
}
\value{
Vector of non-negative integers.
}
\description{
Calculates the number of behaviors that begin during the observation session.
}
\examples{
BS <- r_behavior_stream(n = 5, mu = 3, lambda = 10,
                       F_event = F_exp(), F_interim = F_exp(), stream_length = 100)
event_counting(BS)
}

