% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{FMexample}
\alias{Dunlap}
\alias{FMexample}
\alias{Moes}
\title{Dunlap et al.(1994) and Moes(1998) data}
\format{A data frame with 53940 rows and 10 variables}
\description{
Single case design data from two different studies examining the impact of choice-making on disruptive behavior in academic settings. Thus, baseline is "no choice" and treatment was "choice." Data were extracted from the figures in the publicatons.
}
\details{
For the Dunlap dataset Ahmad and Sven were measured with an active interval of 10s with 5s for recording while Wendall had an active interval length of 15s with no time for recording. Sessions were 15 minutes long and each summary measurement was based on K = 60 intervals.

For the Moes dataset all participants were observed with an active interval of 10s with 5s. Sessions were 20 minutes long and each summary measurement was based on K = 80 intervals.
}
\references{
Dunlap, G., DePerczel, M., Clarke, S., Wilson, D., Wright, S., White, R., & Gomez, A. (1994). Choice making to promote adaptive behavior for students with emotional and behavioral challenges. Journal of Applied Behavior Analysis, 27 (3), 505-518.

Moes, D. R. (1998). Integrating choice-making opportunities within teacher-assigned academic tasks to facilitate the performance of children with autism. Research and Practice for Persons with Severe Disabilities, 23 (4), 319-328.
}
\keyword{datasets}

