% Generated by roxygen2 (4.0.1): do not edit by hand
\name{logRespRatio}
\alias{logRespRatio}
\title{Bias corrected log ratio and variance}
\usage{
logRespRatio(observations, phase, base_level = NULL, bias_correct = TRUE)
}
\arguments{
\item{observations}{vector of observations}

\item{phase}{factor or vector indicating levels of the PIR measurements.}

\item{base_level}{an optional value indicating the name of the baseline level. If nothing is provided, the first level in phase will be treated as the baseline level. Default is \code{NULL}}

\item{bias_correct}{logical value indicating if the bias-corrected log-response ratio should be used. Default is \code{null}}
}
\value{
A list containing two elements. The first element, \code{respRatio}, is the log-response ratio, either corrected or uncorrected depending upon the value of \code{bias_correct}. The second element, \code{variance}, is the variance of the log-response ratio.
}
\description{
Estimates the bias corrected log ratio and the variance of the log ratio
}
\details{
The \code{observations} vector can be in any order corresponding to the factor or vector \code{phase}.
The levels of \code{phase} can be any two levels, such as "A" and "B", "base" and "treat", or "0" and "1".
If there are more than two levels in \code{phase} this function will not work.
By default it is assumed that the first level in \code{phase} is the baseline level,
but the baseline level can be specified via the \code{base_level} variable
}
\examples{
#Get a log ratio and variance estimate for Carl from Moes dataset
data(Moes)
with(subset(Moes, Case == "Carl"),
logRespRatio(observations = outcome, phase = Phase, base_level = "No Choice"))
}

