% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size_estimators.R
\docType{data}
\name{Moes}
\alias{Moes}
\title{Moes(1998) data}
\format{
A data frame with 80 observations on 7 variables

\itemize{
\item \code{Case} The participant for whom the observation took place
\item \code{Phase} The level of the observation ("Choice" vs. "No Choice")
\item \code{Session} The observation session # for each participant
\item \code{outcome} The summary PIR measurement for the observation session
\item \code{active_length} The length of the active observation interval, in seconds
\item \code{rest_length} The length of the recording interval, in seconds
\item \code{intervals} The total number of intervals in the observation session
}

A data frame with 53940 rows and 10 variables
}
\description{
Single-case design data from a study that using partial interval recording (PIR) examining the impact of choice-making in a homework tutoring context on disruptive behavior. In this data "No Choice" is the baseline phase. Data were extracted from the figure in the publication.
}
\references{
Moes, D. R. (1998). Integrating choice-making opportunities within teacher-assigned academic tasks to facilitate the performance of children with autism. Research and Practice for Persons with Severe Disabilities, 23 (4), 319-328.
}
\keyword{datasets}
