% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIR-APP.R
\name{PIR_loglik}
\alias{PIR_loglik}
\title{Calculate log-likelihood}
\usage{
PIR_loglik(param, U, c, d)
}
\arguments{
\item{param}{vector of parameter values logit(prevalence), log(incidence)}

\item{U}{a vector containing interval-level PIR data}

\item{c}{the length of the active interval}

\item{d}{the length of the recording interval

The vector \code{U} should only contain values of 1 or 0.
\code{c} must be a positive value. \code{d} must be some non-negative value. Setting 
\code{d} equal to zero represents a PIR observation where no time was set aside for recording.}
}
\value{
The value of the log-likelihood
}
\description{
Calculates the log-likelihood of within-session PIR data
}
\author{
Daniel Swan <dswan@utexas.edu>
}

