% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulate_behavior_stream.R
\name{r_PIR}
\alias{r_PIR}
\title{Generates random partial interval recording behavior streams}
\usage{
r_PIR(n, mu, lambda, stream_length, F_event, F_interim, interval_length,
  rest_length)
}
\arguments{
\item{n}{number of behavior streams to generate}

\item{mu}{mean event duration}

\item{lambda}{mean interim time}

\item{stream_length}{length of behavior stream}

\item{F_event}{distribution of event durations. Must be of class \code{\link{eq_dist}}.}

\item{F_interim}{distribution of interim times. Must be of class \code{\link{eq_dist}}.}

\item{interval_length}{length of the active observation interval}

\item{rest_length}{length of the recording interval}
}
\description{
Random generation of behavior streams (based on an alternating
renewal process) of a specified length and with specified mean event
durations, mean interim times, event distribution, and interim distribution,
which are then coded as partial interval recording data with given interval length
and rest length.
}
\details{
Generates behavior streams by repeatedly drawing random event durations and
random interim times from the distributions as specified, until the sum of the durations and interim
times exceeds the requested stream length.
}

