% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ARPsimulator.R
\name{ARPsimulator}
\alias{ARPsimulator}
\title{An interactive alternating renewal process simulator}
\usage{
ARPsimulator()
}
\description{
An interactive tool that simulates single-case designs with outcomes measured by systematic direct observation.
The behavioral outcomes are generated from the alternating renewal process model. Both event behaviors and
state behaviors are supported. Event behaviors are generated from a renewal process with gamma-distributed
inter-event times. State behaviors are generated from an alternating poisson process, in which the event
durations and interim times are exponentially distributed. Currently, multiple baseline and treatment reversal
designs are supported.
}

