% Generated by roxygen2 (4.0.1): do not edit by hand
\name{logRespRatio}
\alias{logRespRatio}
\title{Calculate log-response ratio and variance}
\usage{
logRespRatio(observations, phase, base_level, bias_correct = TRUE)
}
\arguments{
\item{observations}{Vector of observations}

\item{phase}{Factor or vector indicating levels of the PIR measurements.}

\item{base_level}{a character string or value indicating the name of the baseline level.}

\item{bias_correct}{Logical value indicating if the bias-corrected log-response ratio should be used. Default is \code{TRUE}}
}
\value{
A list containing two elements.
The first element, \code{lRR}, is the estimated log-response ratio.
The second element, \code{V_lRR}, is the estimated variance of the log-response ratio.
}
\description{
Estimates the bias corrected log-response ratio and the variance of the log-response ratio
}
\details{
The \code{observations} vector can be in any order corresponding to the factor or vector \code{phase}.
The levels of \code{phase} can be any two levels, such as "A" and "B", "base" and "treat", or "0" and "1".
If there are more than two levels in \code{phase} this function will not work.
A value for \code{base_level} must be specified - if it is a chaaracter string it is case sensitive.
}
\examples{
#Get a log ratio and variance estimate for Carl from Moes dataset
data(Moes)
with(subset(Moes, Case == "Carl"),
logRespRatio(observations = outcome, phase = Phase, base_level = "No Choice"))
}
\author{
Daniel Swan <dswan@utexas.edu>
}

