% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PIR_MOM}
\alias{PIR_MOM}
\title{Moment estimator for prevalence and incidence and bootstrap confidence intervals}
\usage{
PIR_MOM(PIR, phase, base_level, intervals, interval_length, rest_length = 0,
  Bootstraps = 2000, conf_level = 0.95, seed = NULL)
}
\arguments{
\item{PIR}{vector of PIR measurements}

\item{phase}{factor or vector indicating levels of the PIR measurements.}

\item{base_level}{a character string or value indicating the name of the baseline level.}

\item{intervals}{the number of intervals in the sample of observations}

\item{interval_length}{the total length of the intervals}

\item{rest_length}{length of the portion of the interval devoted to recording. Default is \code{0}}

\item{Bootstraps}{desired number of bootstrap replicates. Default is \code{2000}}

\item{conf_level}{Desired coverage rate of the calculated confidence interval. Default is \code{.95}.}

\item{seed}{seed value set in order to make bootstrap results reproducible. Default is \code{null}}
}
\value{
A dataframe with three rows corresponding to baseline, treatment, and treatment/baseline and six columns
}
\description{
Estimates prevalance and incidence for two samples as well as the log ratios. Also provides boostrap confidence intervals using the functions \code{\link{r_behavior_stream}} and \code{\link{interval_recording}} to generate the replicates.
}
\details{
The \code{PIR} vector can be in any order corresponding to the factor or vector \code{phase}. The levels of \code{phase} can be any two levels, such as "A" and "B", "base" and "treat", or "0" and "1". If there are more than two levels in \code{phase} this function will not work. A value for \code{base_level} must be specified - if it is a character string it is case sensitive.

\code{intervals}, \code{interval_length}, and \code{rest_length} are all single values that are assumed to be held constant across both samples and all observation sessions. If vectors of values are provided for these variables, it is assumed that the first value in each vector is constant across all observations.

\code{interval_length} This is the length of each individual interval. Sometimes a portion of the interval is set aside for recording purposes. The length of that recording interval is what the value of \code{rest_length} should be set to, although the default assumption is that there is no recording time. The length of the active interval is calculated to be \code{interval_length - rest_length}.

\code{bootstraps} defaults to 2000. This is believed to be a sufficient number. At the default, PIR_MOM takes just under six seconds to run on an Intel Core i5-2410M processor.
}
\examples{
#get estimate and CIs for Carl from the Moes dataset
data(Moes)
with(subset(Moes, Case == "Carl"),
PIR_MOM(PIR = outcome, phase = Phase, intervals = intervals,
interval_length = (active_length + rest_length), rest_length = rest_length,
base_level = "No Choice", seed = 149568373))
}
\author{
Daniel Swan <dswan@utexas.edu>
}

