% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PIR_MOM}
\alias{PIR_MOM}
\title{Moment estimator for prevalence and incidence and bootstrap confidence interval}
\usage{
PIR_MOM(PIR, phase, intervals, interval_length, rest_length = 0,
  first_level_base = TRUE, Bootstraps = 2000, conf_level = 0.95,
  seed = NULL)
}
\arguments{
\item{PIR}{vector of PIR measurements}

\item{phase}{factor or vector indicating levels of the PIR measurements.}

\item{intervals}{the number of intervals in the sample of observations}

\item{interval_length}{the total length of the intervals}

\item{rest_length}{\code{0} (default) or length of the portion of the interval devoted to recording}

\item{first_level_base}{logical. Is the first level in the factor or vector \code{phase} the baseline level?}

\item{Bootstraps}{\code{2000} (default) or the desired number of bootstrap replicates}

\item{conf_level}{\code{.95} (default) or the desired nominal confidence level of the calculated confidence interval}

\item{seed}{\code{NULL} (default) or a seed value set in order to make bootstrap results reproducible}
}
\value{
A list containing four elements
}
\description{
This function provides estimates of prevalance and incidence for two samples as well as the log ratios. It provides boostrap confidence intervals using the functions \code{\link{r_behavior_stream}} and \code{\link{interval_recording}} to generate the replicates.
}
\details{
The \code{PIR} vector can be in any order corresponding to the factor or vector \code{phase}. The levels of \code{phase} can be any two levels, such as "A" and "B", "base" and "treat", or "0" and "1". If there are more than two levels in \code{phase} this function will not work. By default it is assumed that the first level in \code{phase} is the baseline level. This behavior can be changed by \code{first_level_base} to \code{FALSE}.

\code{intervals}, \code{interval_length}, and \code{rest_length} are all single values that are assumed to be held constant across both samples and all observation sessions.

\code{interval_length} This is the length of each individual interval. Sometimes a portion of the interval is set aside for recording purposes. The length of that recording interval is what the value of \code{rest_length} should be set to, although the default assumption is that there is no recording time. The length of the active interval is calculated to be \code{interval_length - rest_length}.

\code{bootstraps} defaults to 2000. This is believed to be a sufficient number. At the default, PIR_MOM takes just under six seconds to run on an Intel Core i5-2410M processor.

\code{seed} defaults to \code{NULL}, but allows for reproducible code.
}

