% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prevalence_bounds}
\alias{prevalence_bounds}
\title{Prevalence bounds and confidence interval}
\usage{
prevalence_bounds(PIR, phase, mu_L, active_length, intervals = NA,
  first_level_base = TRUE, conf_level = 0.95, exponentiate = FALSE)
}
\arguments{
\item{PIR}{vector of PIR measurements}

\item{phase}{factor or vector indicating levels of the PIR measurements.}

\item{mu_L}{the lower limit on the mean event duration}

\item{active_length}{length of the active observation interval}

\item{intervals}{the number of intervals in the sample of observations. Default is \code{NA}.}

\item{first_level_base}{Logical value indicating if the first level of \code{phase} is the baseline level. Default is \code{TRUE}.}

\item{conf_level}{Desired coverage rate of the calculated confidence interval. Default is \code{.95}.}

\item{exponentiate}{Logical value indicating if the log of the bounds and the confidence interval should be exponentiated. Default is \code{FALSE}.}
}
\value{
A list containing four elements
}
\description{
Calculates a bound for the log of the prevalence ratio of two samples (referred to as baseline and treatment)
based on partial interval recording (PIR) data.
}
\details{
The \code{PIR} vector can be in any order corresponding to the factor or vector \code{phase}.
The levels of \code{phase} can be any two levels, such as "A" and "B", "base" and "treat", or "0" and "1".
If there are more than two levels in \code{phase} this function will not work.
By default it is assumed that the first level in \code{phase} is the baseline level,
but this can be changed by \code{first_level_base} to \code{FALSE}.

\code{mu_L} is the lower limit on the mean event durations. This is a single value assumed to hold for both samples of behavior

\code{active_length} This is the total active observation length. If the intervals are 15 seconds long but 5 seconds of each interval is reserved for recording purposes, \code{active_length= 10}. Times are often in seconds, but can be in any time unit.

\code{intervals} is the number of intervals in the observations. This is a single value and is assumed to be constant across both samples and all observations. This value is only relevant if the mean of one of the samples is at the floor or ceiling of 0 or 1. In that case it will be used to truncate the sample mean. If the sample mean is at the floor or ceiling and no value for \code{intervals} is provided, the function will stop.
}
\examples{
#Get an estimate and CI for Carl from Moes dataset
data(FMexample)
with(subset(Moes, Case == "Carl"),
 prevalence_bounds(PIR = outcome, phase = Phase, mu_L = 10, active_length = 10, first_level_base = FALSE))
}

